requirements.txt
pillow
pyTelegramBotAPI
boto3